Reproduce cpprestsdk SEGV in native and Mystikos
The crash happens in epoll_reactor's start_op function.
The function first checks that descriptor_data is not null,
then acquires a lock and proceeds to make use of descriptor_data.

Between the check and the use, another thread could delete the descriptor.
To increase the chances of this happening, put a breakpoint immediately
after the check. When the breakpoint is hit and execution is continued,
the scheduler may schedule other threads instead of the one that just
did the check, increasing the chances of the socket being closed and
the descriptor getting deleted.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>